Update deployment script and add deployment info
